learned props and useState hooks
