+++ index.html
@@ -19,6 +19,6 @@
         crossorigin 
         src="https://unpkg.com/react-dom@18/umd/react-dom.development.js">
     </script>    -->
-    <script type="module" src="./App.js"></script>   
+    <script type="module" src="./src/App.js"></script>   
 </body>
 </html>
