+++ src/App.js
@@ -1,0 +1,26 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import Header from "./components/Header";
+import Body from "./components/Body";
+
+
+// JSX (transpiled before it reaches the JS) - PARCEL - Babel
+
+// JSX => Babel transpiles it to React.createElement => ReactElement-JS Object => HTMLELement(render)
+
+// React Component
+
+
+const AppLayout = () => {
+    return (
+        <div className="app">
+            <Header/>
+            <Body/>
+        </div>
+    )
+}
+
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(<AppLayout/>);
