+++ src/utils/mockData.js
@@ -1,48 +1,3 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import Logo from "./KhaoPio.png";
-
-// JSX (transpiled before it reaches the JS) - PARCEL - Babel
-
-// JSX => Babel transpiles it to React.createElement => ReactElement-JS Object => HTMLELement(render)
-
-// React Component
-
-const Header = () => {
-    return (
-        <div className="header">
-            <div className="logo-container">
-                <img src={Logo} className="logo"/>
-            </div>
-            <div className="nav-items">
-                <ul>
-                    <li>Home</li>
-                    <li>About</li>
-                    <li>Contact Us</li>
-                    <li>Cart</li>
-                </ul>
-            </div>           
-        </div>
-    )
-}
-
-const RestaurantCard = (props) =>{
-    const {resData} = props;
-    const {name, cuisines, avgRating, costForTwo, deliveryTime} = resData?.data;
-    return (
-        <div className="res-card">
-            <img src={"https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/"+resData?.data?.cloudinaryImageId }
-            alt="res-logo" className="res-logo"/>
-
-            <h3>{name}</h3>
-            <h4>{cuisines.join(",")}</h4>
-            <h4>{avgRating}⭐</h4>
-            <h4>₹{costForTwo/100 +" For Two"}</h4>
-            <h4>{deliveryTime +"minutes"}</h4>
-        </div>
-    )
-}
-
 const resList =  [
     {
       type: "restaurant",
@@ -1177,34 +1132,4 @@ const resList =  [
     }
 ];
 
-
-
-const Body = () =>{
-    return (
-        <div className="body">
-            <div className="search">Search</div>
-            <div className="res-container">
-            {
-                resList.map((rest) => (
-                    <RestaurantCard  key= {rest?.data?.id} resData={rest}/>
-                    ))
-            }
-            </div>
-        </div>
-
-    )
-}
-
-const AppLayout = () => {
-    return (
-        <div className="app">
-            <Header/>
-            <Body/>
-        </div>
-    )
-}
-
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-root.render(<AppLayout/>); 
+export default resList;
